Finished SeekVisionTarget command
We updated the SeekVisionTarget command to work with the new distance calculating function. Also, it has come to my attention that within the limelight subsystem we need a final variable for the angle that the limelight is mounted at, which should be accounted for when calculating things such as distance.